requirements.txt
streamlit
boto3
PyPDF2
python-dotenv
botocore
anthropic